refactor: centralize translations
